Pace Supply/Pace Supply/edi_mapping_generator/requirements.txt
PyMuPDF>=1.23.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
httpx>=0.25.0

pyinstaller>=6.0.0
